basic pagination for settings panel, extra mobile features
